Add more files to the os folder that belong there
