+++ lib/src/os/osInterrupts.c
@@ -1,0 +1,8 @@
+/* The comment below is needed for this file to be picked up by generate_ld */
+/* RAM_POS: 0x800D2560 */
+
+#include "macros.h"
+
+//To the best of my knowledge these files can never actually be decompiled
+GLOBAL_ASM("lib/asm/non_matchings/unknown_0D3160/__osDisableInt.s")
+GLOBAL_ASM("lib/asm/non_matchings/unknown_0D3160/__osRestoreInt.s")
