+++ lib/src/os/osViMgr.c
