+++ lib/src/unknown_0D29F0.c
@@ -127,7 +127,7 @@ void osSpTaskLoad(OSTask *intp){
 		;
 }
 
-void osSpTaskStartGo(OSTask *tp){
+void osSpTaskStartGo(OSTask *tp) {
 	while (__osSpDeviceBusy());
 
 	__osSpSetStatus(SP_SET_INTR_BREAK | SP_CLR_SSTEP | SP_CLR_BROKE | SP_CLR_HALT);
